+++ R/idispatch.R
@@ -10,6 +10,7 @@
 #' @param ... a set of `function name`=`interfacer::iface` pairs
 #' @param .default a function to apply in the situation where none of the rules
 #'   can be matched. The default results in an error being thrown.
+#' @param .prune get rid of excess columns that are not in the spec.
 #'
 #' @return the result of dispatching the dataframe to the first function that
 #'   matches the rules in `...`. Matching is permissive in that the test is
@@ -45,14 +46,14 @@
 #'   mean(df$col2)+uplift
 #' }
 #'
-#' # this input matches `i1` and the `extract_mean` call is dispatched
-#' # via `extract_mean.i1`
+#' # this input matches `i2` and the `extract_mean` call is dispatched
+#' # via `extract_mean.i2`
 #' test = tibble::tibble( col2 = 1:10 )
-#' tmp = extract_mean(test, uplift = 50)
+#' tmp = extract_mean(test, uplift = 50, this_env = TRUE)
 #' testthat::expect_equal(tmp, 55.5)
 #'
-#' # this input matches `i2` and the `extract_mean` call is dispatched
-#' # via `extract_mean.i2` and the uplift is not applied
+#' # this input matches `i1` and the `extract_mean` call is dispatched
+#' # via `extract_mean.i1` and the uplift is not applied
 #' test2 = tibble::tibble( col1 = 1:10 )
 #' tmp2 = extract_mean(test2, uplift = 50)
 #' testthat::expect_equal(tmp2, 5.5)
@@ -65,11 +66,30 @@
 #' # generates an error.
 #' test3 = tibble::tibble( wrong_col = 1:10 )
 #' try(extract_mean(test3, uplift = 50))
-idispatch = function(x, ..., .default = NULL) {
-  # have to dispatch using declared params from caller environment
-  env = rlang::caller_env()
+idispatch = function(x, ..., .default = NULL, .prune = FALSE) {
+  # This function redirects program flow based on the format of `x` by calling
+  # one of a set of different functions with similar parameters. We however are
+  # also validating `x` and may make some type changes before dispatching.
+
+  # Firstly we need to find the environment that we are trying to dispatch from
+  # This is the caller environment in which
+  # the function that called the idispatch was called (hence two levels up).
+  # this could be the global environment but may not be.
+  env = rlang::caller_env(2)
+  # We are potentially going to modify it so we copy it first.
   env = rlang::env_clone(env)
+  # The environment (or namespace) in which the function that called idispatch
+  # is defined in: we assume the dispatch targets are also defined in this
+  # enviroment / namespace
+  # It seems functions do not need to be visible in for this code to work
+  # and its hard to test, but running the examples in `demo_ivalidate()` works
+  # from a fresh session.
+  fn_env = rlang::fn_env(rlang::caller_fn())
+  # This is how the call was that triggered the function that triggered this
+  # idispatch call. This is the call we are going to modify:
   call = rlang::caller_call()
+  # This is the dots of the idispatch call. This is a named list of format
+  # specifications where the name defines a function to dispatch to.
   dots = rlang::list2(...)
   if (any(names(dots) == "")) {
     stop("all parameters must be named", call. = FALSE)
@@ -81,42 +101,50 @@ idispatch = function(x, ..., .default = NULL) {
   for (i in seq_along(dots)) {
     fn_name = names(dots)[[i]]
     ifc = dots[[i]]
-    if (!exists(fn_name, mode = "function", envir = env)) {
-      stop("Cannot find dispatch function: ", fn_name, call. = FALSE)
+    if (!exists(fn_name, mode = "function", envir = fn_env)) {
+      stop(
+        "Cannot find dispatch function: ",
+        fn_name,
+        "\n (Maybe it needs to be exported?)",
+        call. = FALSE
+      )
     }
 
-    x2 = try(iconvert(x, ifc, .dname = "nested"), silent = TRUE)
+    dname = deparse(substitute(x))
+    x2 = try(
+      iconvert(x, ifc, .dname = dname, .fname = fn_name, .prune = .prune),
+      silent = TRUE
+    )
 
     if (!inherits(x2, "try-error")) {
-      fn = tryCatch(
-        get(fn_name, mode = "function", envir = env),
-        error = function(e) {
-          stop("could not find function: ", fn_name, call. = FALSE)
-        }
-      )
+      fn_quo = rlang::new_quosure(as.symbol(fn_name), fn_env)
 
       # This has converted the input x into the correct format x2
-      # This is then inserted back into the caller environment under a new name
-      # the original (undispatched) call is modified to the matching function
-      # name and the new variable name.
+      # This is then inserted back into the caller environment
+      # the original (undispatched) call is modified to use the quosure
+      # using the matching function name.
 
       orig_name = format(call[[2]])
-      conv_name = sprintf(".%s.%s", orig_name, fn_name)
-      # TODO: review this
-      # Weird behaviour when a function called with fn(x, y=fn2(x)) becuse
-      # the call becomes fn(x_mod, y=fn2(x)) but x does not exist maybe??
-      old = if (exists(orig_name, envir = env)) {
-        get(orig_name, envir = env)
-      } else {
-        NULL
-      }
-      assign(orig_name, x, envir = env)
-      assign(conv_name, x2, envir = env)
-      call[[1]] = as.symbol(fn_name)
-      call[[2]] = as.symbol(conv_name)
-      tmp = eval(call, envir = env)
-      assign(conv_name, NULL, envir = env)
-      assign(orig_name, old, envir = env)
+      assign(orig_name, x2, envir = env)
+
+      # Alternatively it couldbe inserted under a new namw and the call
+      # modified to use the new variable name but this turns out to be not
+      # needed:
+      # conv_name = sprintf(".%s.%s", orig_name, fn_name)
+      # assign(conv_name, x2, envir = env)
+      # call[[2]] = as.symbol(conv_name)
+
+      # The following can maybe fail if the namespaces of the call are different and
+      # this is hard to test for as it needs to be run in a different package.
+      # call[[1]] = as.symbol(fn_name)
+      # tmp = eval(call, envir = env)
+
+      call[[1]] = fn_quo
+      tmp = rlang::eval_tidy(call, env = env)
+
+      # Do we need to tidy up. I think not as this is a disposable environment
+      # assign(conv_name, NULL, envir = env)
+      # assign(orig_name, old, envir = env)
       return(tmp)
     } else {
       errors = c(errors, fn_name, " - ", as.character(x2))
@@ -138,3 +166,41 @@ idispatch = function(x, ..., .default = NULL) {
     return(eval(call, envir = env))
   }
 }
+
+#
+# # this input matches `i2` and the `extract_mean` call is dispatched
+# # via `extract_mean.i2`
+# test = tibble::tibble( col2 = 1:10 )
+# tmp = extract_mean(test, uplift = 50, this_env = TRUE)
+# quos = eval(expression(rlang::enquos(...)),envir = env)
+# quos
+# quos = eval(expression(rlang::enquos(...)),envir = env)
+# ls(env)
+# ifc
+# fn
+# quos
+# formals(fn)
+# names(formals(fn))[[1]]
+# p1nm = names(formals(fn))[[1]]
+# quos[[p1nm]] = rlang::quo(x2)
+# quos
+# do.call(fn, quos)
+# fn(!!!quos)
+# fn(!!quos)
+# fn(rlang::inject(quos))
+# quos
+# fn
+# x2
+# do.call(fn, rlang::inject(quos)))
+# do.call(fn, rlang::inject(quos))
+# quos
+# ls(env)
+# env$df
+# rlang::fn_env(rlang::caller_fn())
+# ls(rlang::fn_env(rlang::caller_fn()))
+# rlang::caller_fn()
+# rlang::caller_fn(2)
+# rlang::caller_fn()
+# rlang::fn_env(rlang::caller_fn())
+# fnenv = rlang::fn_env(rlang::caller_fn())
+# fn = get(fn_name, mode = "function", envir = fnenv)